+++ src/itest/java/com/iexec/worker/compute/post/IntegrationTests.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2022-2023 IEXEC BLOCKCHAIN TECH
+ * Copyright 2022-2025 IEXEC BLOCKCHAIN TECH
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -65,10 +65,10 @@ class IntegrationTests {
     @Test
     void shouldHandleCallback(EnvironmentVariables environment) {
         final String taskId = "0x0000000000000000000000000000000000000000000000000000000000000001";
-        environment.set("RESULT_TASK_ID", taskId);
+        environment.set("IEXEC_TASK_ID", taskId);
         environment.set("RESULT_STORAGE_CALLBACK", "yes");
-        environment.set("RESULT_SIGN_WORKER_ADDRESS", "0x0000000000000000000000000000000000000002");
-        environment.set("RESULT_SIGN_TEE_CHALLENGE_PRIVATE_KEY", "0x000000000000000000000000000000000000000000000000000000000000003");
+        environment.set("SIGN_WORKER_ADDRESS", "0x0000000000000000000000000000000000000002");
+        environment.set("SIGN_TEE_CHALLENGE_PRIVATE_KEY", "0x000000000000000000000000000000000000000000000000000000000000003");
         environment.set("WORKER_HOST", worker.getServiceHost(WORKER_SERVICE_NAME, WORKER_SERVICE_PORT) + ":" + worker.getServicePort(WORKER_SERVICE_NAME, WORKER_SERVICE_PORT));
 
         try (MockedStatic<IexecFileHelper> iexecFileHelper = Mockito.mockStatic(IexecFileHelper.class)) {